pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "autogen-agentchat>=0.7.4",
    "autogen-ext[openai]>=0.7.4",
    "flask>=3.0.0",
    "flask-cors>=6.0.1",
    "gunicorn>=23.0.0",
    "python-dotenv>=1.0.0",
]
